\name{fluEvidenceSynthesis-package}
\alias{fluEvidenceSynthesis-package}
\alias{fluEvidenceSynthesis}
\docType{package}
\title{
\packageTitle{fluEvidenceSynthesis}
}
\description{
\packageDescription{fluEvidenceSynthesis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fluEvidenceSynthesis}
\packageIndices{fluEvidenceSynthesis}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{fluEvidenceSynthesis}

Maintainer: \packageMaintainer{fluEvidenceSynthesis}
}
\references{
Baguelin, Marc, Stefan Flasche, Anton Camacho, Nikolaos Demiris, Elizabeth Miller, and W. John Edmunds. \emph{Assessing Optimal Target Populations for Influenza Vaccination Programmes: An Evidence Synthesis and Modelling Study}. PLoS Med 10, no. 10 (8 October 2013): e1001527. doi:10.1371/journal.pmed.1001527.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
